requirements.txt
# Assuming annotated-types is incorrect or not needed, it's removed.
# If it's a necessary package, please replace it with the correct name and version.

anyio==4.2.0
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
et-xmlfile==1.1.0
Flask==2.1.1
Flask-Cors==3.0.10
gunicorn==20.1.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.3
openai==1.7.2
openpyxl==3.1.2
packaging==23.2
pandas==2.1.4
pandas-stubs==2.1.4.231227
pip==23.3.2
pydantic==2.5.3
pydantic_core==2.14.6
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2023.3.post1
requests==2.31.0
setuptools==69.0.3
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
types-pytz==2023.3.1.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
Werkzeug==2.1.2
